+++ framework/packages/abstract-core/src/native/module_factory.rs
@@ -5,13 +5,8 @@
 //! ## Description  
 //! This contract is instantiated by Abstract and only used internally. Adding or upgrading modules is done using the [`crate::manager::ExecuteMsg`] endpoint.  
 pub mod state {
-    use std::collections::VecDeque;
 
-    use crate::{
-        manager::RegisterModuleData,
-        objects::module::{Module, ModuleInfo},
-        version_control::AccountBase,
-    };
+    use crate::{objects::module::ModuleInfo, version_control::AccountBase};
     use cosmwasm_std::{Addr, Binary};
     use cw_storage_plus::{Item, Map};
     use schemars::JsonSchema;
@@ -26,8 +21,6 @@ pub mod state {
     #[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
     pub struct Context {
         pub account_base: AccountBase,
-        pub modules: VecDeque<Module>,
-        pub modules_to_register: Vec<RegisterModuleData>,
     }
 
     pub const CONFIG: Item<Config> = Item::new("\u{0}{5}config");
@@ -35,11 +28,7 @@ pub mod state {
     pub const MODULE_INIT_BINARIES: Map<&ModuleInfo, Binary> = Map::new("module_init_binaries");
 }
 
-use crate::{
-    manager::RegisterModuleData,
-    objects::module::{Module, ModuleInfo},
-    version_control::AccountBase,
-};
+use crate::{objects::module::ModuleInfo, version_control::AccountBase};
 use cosmwasm_schema::QueryResponses;
 use cosmwasm_std::{Addr, Binary, Coin};
 
@@ -63,22 +52,25 @@ pub enum ExecuteMsg {
         version_control_address: Option<String>,
     },
     /// Install modules
-    InstallModules { modules: Vec<ModuleInstallConfig> },
+    InstallModules {
+        modules: Vec<FactoryModuleInstallConfig>,
+        salt: Binary,
+    },
     UpdateFactoryBinaryMsgs {
         to_add: Vec<(ModuleInfo, Binary)>,
         to_remove: Vec<ModuleInfo>,
     },
 }
 
-/// Module info and init message
+/// Module info, init message and salt
 #[non_exhaustive]
 #[cosmwasm_schema::cw_serde]
-pub struct ModuleInstallConfig {
+pub struct FactoryModuleInstallConfig {
     pub module: ModuleInfo,
     pub init_msg: Option<Binary>,
 }
 
-impl ModuleInstallConfig {
+impl FactoryModuleInstallConfig {
     pub fn new(module: ModuleInfo, init_msg: Option<Binary>) -> Self {
         Self { module, init_msg }
     }
@@ -114,8 +106,6 @@ pub struct ConfigResponse {
 #[cosmwasm_schema::cw_serde]
 pub struct ContextResponse {
     pub account_base: AccountBase,
-    pub modules: Vec<Module>,
-    pub modules_to_register: Vec<RegisterModuleData>,
 }
 
 #[cosmwasm_schema::cw_serde]