+++ framework/packages/abstract-interface/src/native/module_factory.rs
@@ -17,8 +17,7 @@ impl<Chain: CwEnv> Uploadable for ModuleFactory<Chain> {
                 ::module_factory::contract::instantiate,
                 ::module_factory::contract::query,
             )
-            .with_migrate(::module_factory::contract::migrate)
-            .with_reply(::module_factory::contract::reply),
+            .with_migrate(::module_factory::contract::migrate),
         )
     }
     fn wasm(&self) -> WasmPath {